+++ core/router/party.go
@@ -27,6 +27,10 @@ type Party interface {
 	//
 	// It returns the same `APIBuilder` featured with Dependency Injection.
 	ConfigureContainer(builder ...func(*APIContainer)) *APIContainer
+	// EnsureStaticBindings panics on struct handler (controller)
+	// if at least one input binding depends on the request and not in a static structure.
+	// Should be called before `RegisterDependency`.
+	EnsureStaticBindings() Party
 	// RegisterDependency calls the `ConfigureContainer.RegisterDependency` method
 	// with the provided value(s). See `HandleFunc` and `PartyConfigure` methods too.
 	RegisterDependency(dependencies ...interface{})